@@ -14,79 +14,78 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.8.0 <3.0.0"
+  sdk: ">=2.15.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
   # Dialog
-  awesome_dialog: ^1.1.2
+  awesome_dialog: ^2.1.2
 
   # Loading
-  loading_overlay: ^0.2.1
+  loading_overlay: ^0.3.0
 
   # Graph
-  fl_chart: ^0.9.4
+  fl_chart: ^0.46.0
 
   # Navbar
-  bubble_bottom_bar: ^1.2.0
-  curved_navigation_bar: ^0.3.2
+  bubble_bottom_bar: ^2.0.0
+  curved_navigation_bar: ^1.0.3
 
   # Firebase
   firebase_core: ^1.0.1
-  firebase_auth: ^1.0.1
-  firebase_storage: ^8.0.0
-  cloud_firestore: ^1.0.1
-  google_sign_in: ^4.5.3
+  firebase_auth: ^3.3.10
+  firebase_storage: ^10.2.9
+  cloud_firestore: ^3.1.10
+  google_sign_in: ^5.2.4
 
   # Twitter
   flutter_twitter: ^1.1.3
 
   # Sign In With Apple
-  apple_sign_in: ^0.1.0
-  sign_in_with_apple: ^2.5.2
+  sign_in_with_apple: ^3.3.0
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.2
+  cupertino_icons: ^1.0.4
 
   # State
-  state_notifier: ^0.5.0
-  flutter_state_notifier: ^0.4.2
-  flutter_riverpod: ^0.11.0
+  state_notifier: ^0.7.2+1
+  flutter_state_notifier: ^0.7.1
   freezed_annotation:
+  hooks_riverpod: ^1.0.3
 
   # Icon
-  font_awesome_flutter: 8.8.1
+  font_awesome_flutter: ^9.2.0
 
   # Button
-  rounded_loading_button: ^1.0.11
+  rounded_loading_button: ^2.1.0
   floating_action_bubble: ^1.1.2
 
   # Form
-  email_validator: ^1.0.0
+  email_validator: ^2.0.1
   flutter_datetime_picker: ^1.5.0
 
   # Introduction
-  flutter_overboard: ^1.1.5
+  flutter_overboard: ^3.1.1
 
   # Image
-  image_picker: ^0.7.2+1
+  image_picker: ^0.8.4+11
   image_cropper: ^1.3.1
-  flutter_image_compress: ^0.7.0
+  flutter_image_compress: ^1.1.0
   path_provider: ^2.0.1
-  cached_network_image: ^2.5.0
+  cached_network_image: ^3.2.0
 
   # Data
   intl: ^0.17.0
-  uuid: ^2.2.2
+  uuid: ^3.0.6
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
   build_runner:
-  flutter_launcher_icons: ^0.7.5
+  flutter_launcher_icons: ^0.9.2
   freezed:
   json_serializable:
 
