@@ -1,13 +1,12 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/all.dart';
-import 'package:freezed_annotation/freezed_annotation.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:livle/repositories/artist_list.dart';
 import 'package:livle/repositories/money_list.dart';
 import 'package:livle/repositories/schedule_list.dart';
 
 final ChangeNotifierProvider<ArtistViewModel> artistViewModelNotifierProvider = ChangeNotifierProvider<ArtistViewModel>(
-  (ProviderReference ref) => ArtistViewModel(
+  (ChangeNotifierProviderRef<ArtistViewModel> ref) => ArtistViewModel(
     scheduleRepository: ref.watch(scheduleRepositoryProvider),
     artistRepository: ref.watch(artistsRepositoryProvider),
     moneyRepository: ref.watch(moneyRepositoryProvider),
@@ -16,9 +15,9 @@ final ChangeNotifierProvider<ArtistViewModel> artistViewModelNotifierProvider = 
 
 class ArtistViewModel extends ChangeNotifier {
   ArtistViewModel({
-    @required ArtistRepository artistRepository,
-    @required ScheduleRepository scheduleRepository,
-    @required MoneyRepository moneyRepository,
+    required ArtistRepository artistRepository,
+    required ScheduleRepository scheduleRepository,
+    required MoneyRepository moneyRepository,
   })  : _scheduleRepository = scheduleRepository,
         _artistRepository = artistRepository,
         _moneyRepository = moneyRepository;
