@@ -1,12 +1,11 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/all.dart';
-import 'package:freezed_annotation/freezed_annotation.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:livle/repositories/artist_list.dart';
 import 'package:livle/repositories/schedule_list.dart';
 
 final ChangeNotifierProvider<ScheduleViewModel> scheduleViewModelNotifierProvider = ChangeNotifierProvider<ScheduleViewModel>(
-  (ProviderReference ref) => ScheduleViewModel(
+  (ChangeNotifierProviderRef<ScheduleViewModel> ref) => ScheduleViewModel(
     scheduleRepository: ref.watch(scheduleRepositoryProvider),
     artistRepository: ref.watch(artistsRepositoryProvider),
   ),
@@ -14,8 +13,8 @@ final ChangeNotifierProvider<ScheduleViewModel> scheduleViewModelNotifierProvide
 
 class ScheduleViewModel extends ChangeNotifier {
   ScheduleViewModel({
-    @required ScheduleRepository scheduleRepository,
-    @required ArtistRepository artistRepository,
+    required ScheduleRepository scheduleRepository,
+    required ArtistRepository artistRepository,
   })  : _scheduleRepository = scheduleRepository,
         _artistRepository = artistRepository;
 
