@@ -1,13 +1,13 @@
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:firebase_storage/firebase_storage.dart';
 import 'package:flutter/cupertino.dart';
-import 'package:flutter_riverpod/all.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:livle/providers/firebase_providers.dart';
 import 'package:livle/repositories/user.dart';
 import 'package:livle/services/auth_service.dart';
 
 final ChangeNotifierProvider<UserViewModel> userViewModelNotifierProvider = ChangeNotifierProvider<UserViewModel>(
-  (ProviderReference ref) => UserViewModel(
+  (ChangeNotifierProviderRef<UserViewModel> ref) => UserViewModel(
     userRepository: ref.watch(originUserProvider),
     storage: ref.watch(firebaseStorageProvider),
   ),
@@ -15,8 +15,8 @@ final ChangeNotifierProvider<UserViewModel> userViewModelNotifierProvider = Chan
 
 class UserViewModel extends ChangeNotifier {
   UserViewModel({
-    @required OriginUserRepository userRepository,
-    @required FirebaseStorage storage,
+    required OriginUserRepository userRepository,
+    required FirebaseStorage storage,
   })  : _userRepository = userRepository,
         _storage = storage;
 
@@ -29,7 +29,7 @@ class UserViewModel extends ChangeNotifier {
 
   Future<void> downloadUserIcon() async {
     final AuthService _authService = AuthService();
-    final String _uid = _authService.fetchCurrentUser().uid;
+    final String _uid = _authService.fetchCurrentUser()?.uid ?? '';
     final OriginUser _originUser = userRepository.originUser;
     if (_originUser.pickedImage) {
       final Reference _storageRef = _storage.ref().child('user_icons/$_uid.jpg');
