@@ -1,5 +1,5 @@
 import 'package:flutter/foundation.dart';
-import 'package:flutter_riverpod/all.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:livle/providers/firebase_providers.dart';
 import 'package:livle/source/interfaces/user.dart';
@@ -9,8 +9,8 @@ part 'user.freezed.dart';
 
 class OriginUserRepository extends ChangeNotifier {
   OriginUserRepository({
-    @required this.source,
-    @required OriginUser originUser,
+    required this.source,
+    required OriginUser originUser,
   }) : _originUser = originUser;
 
   IUserDataSource source;
@@ -26,8 +26,8 @@ class OriginUserRepository extends ChangeNotifier {
 @freezed
 abstract class OriginUser with _$OriginUser {
   const factory OriginUser({
-    @required String id,
-    @required String name,
+    required String id,
+    required String name,
     @Default('') String description,
     @Default('images/default_user_icon.png') String iconImagePath,
     @Default(false) bool pickedImage,
@@ -37,7 +37,7 @@ abstract class OriginUser with _$OriginUser {
 }
 
 final Provider<OriginUserRepository> originUserProvider = Provider<OriginUserRepository>(
-  (ProviderReference ref) => OriginUserRepository(
+  (ProviderRef<OriginUserRepository> ref) => OriginUserRepository(
     source: UserDataSource(storage: ref.read(firebaseStorageProvider)),
     originUser: const OriginUser(
       id: '',
