@@ -6,11 +6,11 @@ import 'package:livle/source/interfaces/money.dart';
 
 class MoneyDataSource implements IMoneyDataSource {
   @override
-  Future<Map<String, dynamic>> fetch() async {
+  Future<Map<String, dynamic>?> fetch() async {
     final AuthService _authService = AuthService();
-    Map<String, dynamic> data;
-    final String _uid = _authService.fetchCurrentUser().uid;
-    await FirebaseFirestore.instance.collection('money').doc(_uid).get().then((DocumentSnapshot value) {
+    Map<String, dynamic>? data;
+    final String _uid = _authService.fetchCurrentUser()?.uid ?? '';
+    await FirebaseFirestore.instance.collection('money').doc(_uid).get().then((DocumentSnapshot<Map<String, dynamic>?> value) {
       if (value.exists) {
         data = value.data();
       }
@@ -21,8 +21,8 @@ class MoneyDataSource implements IMoneyDataSource {
   @override
   Future<bool> init() async {
     final AuthService _auth = AuthService();
-    final User firebaseUser = _auth.fetchCurrentUser();
-    final String uid = firebaseUser.uid;
+    final User? firebaseUser = _auth.fetchCurrentUser();
+    final String uid = firebaseUser?.uid ?? '';
     await FirebaseFirestore.instance.collection('money').doc(uid).set(<String, dynamic>{
       'spendings': <Map<String, dynamic>>[],
     });
@@ -33,8 +33,8 @@ class MoneyDataSource implements IMoneyDataSource {
   @override
   Future<bool> add(Money money) async {
     final AuthService _auth = AuthService();
-    final User firebaseUser = _auth.fetchCurrentUser();
-    final String uid = firebaseUser.uid;
+    final User? firebaseUser = _auth.fetchCurrentUser();
+    final String uid = firebaseUser?.uid ?? '';
     await FirebaseFirestore.instance.collection('money').doc(uid).update(<String, dynamic>{
       'spendings': FieldValue.arrayUnion(<Map<String, dynamic>>[
         <String, dynamic>{
