@@ -1,13 +1,13 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/all.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:livle/providers/view_model/user.dart';
 
-class HeaderUserIcon extends ConsumerWidget {
-  const HeaderUserIcon({Key key}) : super(key: key);
+class HeaderUserIcon extends HookConsumerWidget {
+  const HeaderUserIcon({Key? key}) : super(key: key);
 
   @override
-  Widget build(BuildContext context, ScopedReader watch) {
-    final UserViewModel _userViewModel = watch(userViewModelNotifierProvider);
+  Widget build(BuildContext context, WidgetRef ref) {
+    final UserViewModel _userViewModel = ref.watch(userViewModelNotifierProvider);
     // _userViewModel.downloadUserIcon();
     return FutureBuilder<void>(
       future: _userViewModel.downloadUserIcon(),
