@@ -1,16 +1,16 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/all.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:livle/providers/view_model/money.dart';
 import 'package:livle/view/components/money/money_list.dart';
 import 'package:livle/view/components/money/pie_chart.dart';
 
-class MoneyGraphView extends ConsumerWidget {
-  const MoneyGraphView({Key key}) : super(key: key);
+class MoneyGraphView extends HookConsumerWidget {
+  const MoneyGraphView({Key? key}) : super(key: key);
 
   @override
-  Widget build(BuildContext context, ScopedReader watch) {
-    final MoneyViewModel _moneyViewModel = watch(moneyViewModelNotifierProvider);
+  Widget build(BuildContext context, WidgetRef ref) {
+    final MoneyViewModel _moneyViewModel = ref.watch(moneyViewModelNotifierProvider);
     return Stack(
       children: <Widget>[
         Container(
