@@ -1,16 +1,16 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/all.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:intl/date_symbol_data_local.dart';
 import 'package:intl/intl.dart';
 import 'package:livle/providers/view_model/money.dart';
 import 'package:livle/repositories/money.dart';
 
-class MoneyDetailListView extends ConsumerWidget {
-  const MoneyDetailListView({Key key}) : super(key: key);
+class MoneyDetailListView extends HookConsumerWidget {
+  const MoneyDetailListView({Key? key}) : super(key: key);
 
   @override
-  Widget build(BuildContext context, ScopedReader watch) {
-    final MoneyViewModel _moneyViewModel = watch(moneyViewModelNotifierProvider);
+  Widget build(BuildContext context, WidgetRef ref) {
+    final MoneyViewModel _moneyViewModel = ref.watch(moneyViewModelNotifierProvider);
     initializeDateFormatting('ja_JP');
     final DateFormat _formatter = DateFormat('yyyy/MM/dd（E）', 'ja_JP');
     return ListView.builder(
