@@ -4,14 +4,14 @@ import 'package:livle/view/components/money/graph_view.dart';
 import 'package:livle/view/components/money/list_view.dart';
 
 class MoneyTabBarView extends StatefulWidget {
-  const MoneyTabBarView({Key key}) : super(key: key);
+  const MoneyTabBarView({Key? key}) : super(key: key);
 
   @override
   _MoneyTabBarViewState createState() => _MoneyTabBarViewState();
 }
 
 class _MoneyTabBarViewState extends State<MoneyTabBarView> with SingleTickerProviderStateMixin {
-  TabController _tabController;
+  late TabController _tabController;
 
   final List<Widget> _tabViews = <Widget>[
     const MoneyGraphView(),
