@@ -1,13 +1,13 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:livle/providers/view_model/money_register.dart';
 
-class MoneyRegisterArtistInput extends ConsumerWidget {
-  const MoneyRegisterArtistInput({Key key}) : super(key: key);
+class MoneyRegisterArtistInput extends HookConsumerWidget {
+  const MoneyRegisterArtistInput({Key? key}) : super(key: key);
 
   @override
-  Widget build(BuildContext context, ScopedReader watch) {
-    final MoneyRegisterViewModel _moneyRegisterViewModel = watch(moneyRegisterViewModelNotifierProvider);
+  Widget build(BuildContext context, WidgetRef ref) {
+    final MoneyRegisterViewModel _moneyRegisterViewModel = ref.watch(moneyRegisterViewModelNotifierProvider);
     return DropdownButtonFormField<String>(
       decoration: const InputDecoration(
         filled: true,
@@ -42,7 +42,7 @@ class MoneyRegisterArtistInput extends ConsumerWidget {
         ),
       ),
       items: _moneyRegisterViewModel.artistDropDownList,
-      onChanged: (String value) => _moneyRegisterViewModel.artistId = value,
+      onChanged: (String? value) => _moneyRegisterViewModel.artistId = value ?? '',
     );
   }
 }
