@@ -1,38 +1,32 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/all.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
-import 'package:livle/providers/view_model/artist.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:livle/providers/firestore_providers.dart';
+import 'package:livle/providers/view_model/artist.dart';
 import 'package:livle/repositories/artist_list.dart';
 import 'package:livle/repositories/money_list.dart';
 import 'package:livle/repositories/schedule_list.dart';
 import 'package:livle/view/components/artist/artist_grid.dart';
 import 'package:livle/view/components/common/loading_indicator.dart';
 
-class ArtistPage extends ConsumerWidget {
-  const ArtistPage({Key key}) : super(key: key);
+class ArtistPage extends HookConsumerWidget {
+  const ArtistPage({Key? key}) : super(key: key);
 
   @override
-  Widget build(BuildContext context, ScopedReader watch) {
-    final ArtistViewModel _artistViewModel = watch(artistViewModelNotifierProvider);
-    final AsyncValue<Map<String, dynamic>> _scheduleStreamProvider = watch(scheduleStreamProvider);
-    final AsyncValue<Map<String, dynamic>> _artistsStreamProvider = watch(artistsStreamProvider);
-    final AsyncValue<Map<String, dynamic>> _moneyStreamProvider = watch(moneyStreamProvider);
+  Widget build(BuildContext context, WidgetRef ref) {
+    final ArtistViewModel _artistViewModel = ref.watch(artistViewModelNotifierProvider);
+    final AsyncValue<Map<String, dynamic>?> _scheduleStreamProvider = ref.watch(scheduleStreamProvider);
+    final AsyncValue<Map<String, dynamic>?> _artistsStreamProvider = ref.watch(artistsStreamProvider);
+    final AsyncValue<Map<String, dynamic>?> _moneyStreamProvider = ref.watch(moneyStreamProvider);
     return _scheduleStreamProvider.when(
-      data: (Map<String, dynamic> data) {
-        if (data != null) {
-          _artistViewModel.scheduleRepository.scheduleList = ScheduleList.fromJson(data);
-        }
+      data: (Map<String, dynamic>? data) {
+          _artistViewModel.scheduleRepository.scheduleList = ScheduleList.fromJson(data ?? <String, dynamic>{});
         return _artistsStreamProvider.when(
-          data: (Map<String, dynamic> data) {
-            if (data != null) {
-              _artistViewModel.artistRepository.artistList = ArtistList.fromJson(data);
-            }
+          data: (Map<String, dynamic>? data) {
+              _artistViewModel.artistRepository.artistList = ArtistList.fromJson(data ?? <String, dynamic>{});
             return _moneyStreamProvider.when(
-              data: (Map<String, dynamic> data) {
-                if (data != null) {
-                  _artistViewModel.moneyRepository.moneyList = MoneyList.fromJson(data);
-                }
+              data: (Map<String, dynamic>? data) {
+                  _artistViewModel.moneyRepository.moneyList = MoneyList.fromJson(data ?? <String, dynamic>{});
                 if (_artistViewModel.artistRepository.artistList.artists.isNotEmpty) {
                   return Container(
                     color: Colors.white,
