@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 // import 'package:url_launcher/url_launcher.dart';
 
 class CommunityPage extends StatelessWidget {
-  const CommunityPage({Key key}) : super(key: key);
+  const CommunityPage({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
