@@ -2,11 +2,11 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 
 class Schedule {
   Schedule({
-    this.id,
-    this.title,
-    this.date,
-    this.memo,
-    this.artistId,
+    required this.id,
+    required this.title,
+    required this.date,
+    required this.memo,
+    required this.artistId,
   });
 
   String id;
