@@ -1,11 +1,9 @@
 // ignore: avoid_classes_with_only_static_members
 class OriginValidators {
-  static final Map<String, String> regs = <String, String>{
-    'userId': r'^[a-zA-Z0-9_]+$',
-  };
+  static String userIdRegs = r'^[a-zA-Z0-9_]+$';
 
-  static String userId(String value, String maxLength) {
-    if (!_validator(value, regs['userId'])) {
+  static String? userId(String value, String maxLength) {
+    if (!_validator(value, userIdRegs)) {
       return '半角英数字とアンダーバー以外使用できません。';
     } else if (value.length > int.parse(maxLength) || value.isEmpty) {
       return '1文字以上$maxLength文字以内で入力してください';
@@ -13,28 +11,28 @@ class OriginValidators {
     return null;
   }
 
-  static String userName(String value, String maxLength) {
+  static String? userName(String value, String maxLength) {
     if (value.length > int.parse(maxLength) || value.isEmpty) {
       return '1文字以上$maxLength文字以内で入力してください';
     }
     return null;
   }
 
-  static String userDesc(String value, String maxLength) {
+  static String? userDesc(String value, String maxLength) {
     if (value.length > int.parse(maxLength)) {
       return '$maxLength文字以内で入力してください';
     }
     return null;
   }
 
-  static String moneyTitle(String value, String maxLength) {
+  static String? moneyTitle(String value, String maxLength) {
     if (value.length > int.parse(maxLength) || value.isEmpty) {
       return '1文字以上$maxLength文字以内で入力してください';
     }
     return null;
   }
 
-  static String moneyAmount(String value) {
+  static String? moneyAmount(String value) {
     if (value.isEmpty) {
       return '入力必須です';
     }
@@ -46,11 +44,11 @@ class OriginValidators {
     return null;
   }
 
-  static String moneyDate(String value) {
+  static String? moneyDate(String value) {
     return null;
   }
 
-  static String moneyMemo(String value, String maxLength) {
+  static String? moneyMemo(String value, String maxLength) {
     if (value.length > int.parse(maxLength)) {
       return '$maxLength文字以内で入力してください';
     }
