@@ -1,7 +1,7 @@
 import 'package:livle/repositories/artist.dart';
 
 abstract class IArtistDataSource {
-  Future<Map<String, dynamic>> fetch();
+  Future<Map<String, dynamic>?> fetch();
   Future<bool> add(Artist money);
   Future<bool> delete(Artist money);
   Future<bool> init();
