@@ -1,16 +1,16 @@
 import 'package:curved_navigation_bar/curved_navigation_bar.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:livle/providers/navigation.dart';
 
-class BottomBar extends ConsumerWidget {
+class BottomBar extends HookConsumerWidget {
   @override
-  Widget build(BuildContext context, ScopedReader watch) {
-    final StateController<int> _pageIndex = watch(pageIndexProvider);
-    final StateController<GlobalKey<CurvedNavigationBarState>> _bottomNavigationKey = watch(bottomNavigationKeyProvider);
+  Widget build(BuildContext context, WidgetRef ref) {
+    final StateController<int> _pageIndex = ref.watch(pageIndexProvider.state);
+    final GlobalKey<CurvedNavigationBarState> _bottomNavigationKey = ref.watch(bottomNavigationKeyProvider);
     return CurvedNavigationBar(
-      key: _bottomNavigationKey.state,
+      key: _bottomNavigationKey,
       index: 0,
       height: 60.0,
       items: const <Widget>[
